[BOD-320] Move Button components to folder & rename CustomProps i
