+++ src/atoms/Alert.tsx
@@ -1,4 +1,4 @@
-import { IconButton, IconButtonProps } from './IconButton'
+import { IconButton, IconButtonProps } from './Button'
 import { Icon } from './Icon'
 import './Alert.scss'
 import { classNames } from '../utils/classNames'
