+++ src/atoms/Button/FloatingButton.scss
@@ -1,5 +1,5 @@
-@use '../settings/color_alias';
-@use '../settings/config';
+@use '../../settings/color_alias';
+@use '../../settings/config';
 
 .floating-button {
   display: inline-flex;
