+++ src/atoms/Button/IconButton.scss
@@ -1,5 +1,5 @@
-@use '../settings/color_alias';
-@use '../settings/config';
+@use '../../settings/color_alias';
+@use '../../settings/config';
 
 .icon-button {
   border: none;
