+++ src/atoms/Button/index.ts
@@ -1,0 +1,9 @@
+import { Button } from './Button'
+import { IconButton } from './IconButton'
+import { FloatingButton } from './FloatingButton'
+
+export type { ButtonProps } from './Button'
+export type { IconButtonProps } from './IconButton'
+export type { FloatingButtonProps } from './FloatingButton'
+
+export { Button, IconButton, FloatingButton }
