+++ src/stories/Changelog.stories.mdx
@@ -6,6 +6,9 @@ import { Meta } from "@storybook/addon-docs";
 ## 0.15.0
 - Added FloatingButton component to Storybook
 
+BREAKING CHANGES
+- All button components have been moved to a common folder
+
 ## 0.14.1
 - Added MenuLink and MenuDropdown components to Storybook and remove MenuOption component.
